fix(ai): no more unnecessary re-renders in AIConversation
